requirements.txt
# Flask Framework
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
Werkzeug==3.0.1

# Document Processing
PyPDF2==3.0.1
pdfplumber==0.10.3
python-docx==1.1.0

# LLM Integration
ollama==0.1.7
google-generativeai==0.3.2

# Data Processing (compatible versions)
pandas>=2.2.0
numpy>=1.26.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0
python-json-logger==2.0.7
python-dateutil==2.8.2